2.3.9
git-svn-id: https://plugins.svn.wordpress.org/feedweb/trunk@839957 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ feedweb_util.php
@@ -547,6 +547,10 @@ function CheckServiceAvailability()
 	if ($dom->loadXML($response['body']) == true)
 		if ($dom->documentElement->tagName == "BANNER")
 		{
+			$error = $dom->documentElement->getAttribute("error");
+			if ($error != null && $error != "")
+				return $error;
+			
 			$query = "DELETE FROM $wpdb->postmeta WHERE post_id=0 AND meta_key='feedweb_last_access'";
 			$wpdb->query($query);
